removed methods that are no longer used, so that non-windows-compatible packages can be removed
